pyproject.toml
[project]
name = "fastnotes"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.3"
dependencies = [
    "alembic>=1.15.2",
    "asyncpg>=0.30.0",
    "fastapi[standard]>=0.115.12",
    "psycopg2>=2.9.10",
    "pydantic-settings>=2.9.1",
    "sqlalchemy[asyncio]>=2.0.41",
]
